Update README.md
Follows the bizarre things that happen when a third-rate lawyer coincidentally takes charge of a murder case and digs deeper into the hidden truth. It is a record of the battles faced by a lawyer of the lower middle class who must become evils than the villains in order to protect himself and his family and punish the bad guys in a world full of conspiracies and greed over power and capital. Park Chang Ho, a lawyer who was never successful to begin with and eventually ends up reaching a win percentage of only 10 percent. Those around him call him a big mouth for talking grandiosely but showing no substance. He is so badly off that he can’t afford to pay the monthly rent for his office nor pay his office manager, who is also his father-in-law, on time. Go Mi Ho who is Park Chang Ho’s wife and the person who made him a lawyer. She was graceful and reserved during her college days, but she has now become extremely bold and foul-mouthed.